requirements.txt
asgiref==3.3.4
Django==3.2
django-filter==2.4.0
django-jet==1.0.8
django-model-utils==4.1.1
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
